renaming class wrapper
