+++ packages/crypto/src/node/cryptoApiWrapper.ts
@@ -3,7 +3,7 @@ import type { CryptoApi } from '../types';
 import { bufferFromString, stringFromBuffer, decrypt, encrypt, keyFromPassword } from './aes-ctr';
 import { randomBytes } from './randomBytes';
 
-export class CryptoWrapper implements CryptoApi {
+export class CryptoApiWrapper implements CryptoApi {
   bufferFromString = bufferFromString;
   stringFromBuffer = stringFromBuffer;
   decrypt = decrypt;
@@ -12,4 +12,4 @@ export class CryptoWrapper implements CryptoApi {
   randomBytes = randomBytes;
 }
 
-export default new CryptoWrapper();
+export default new CryptoApiWrapper();
